Add means to split monticello version files
The version files for monticell (located in monticello.meta for cypress)
are rather huge and, being one long string, not really suited for VCS/SCM
like git, svn, or hg.

With this feature, the version file is split into a 'verison.d' directory
containing one version file per ancestor/stepChild. The original version
file simply includes the UUID of the current version, to be retrieved from
the dir.